+++ src/steps/Refresh.js
@@ -15,10 +15,12 @@ class Refresh extends Step {
     new Terraform().exe({
       args   : { refresh: constants.SHELLARG_PREPEND_AS_IS },
       runtime: this.runtime,
-      mode   : 'silent',
     }, (err, stdout) => {
       if (err) {
         if (`${err.details}`.match(/when there is existing state/)) {
+          debug('Ignoring refresh error about missing statefile')
+          return cb(null)
+        } else if (`${err.details}`.match(/non-existent state file/)) {
           debug('Ignoring refresh error about missing statefile')
           return cb(null)
         } else {